vision-inference-service/requirements.txt
fastapi
uvicorn
transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
Pillow
python-multipart
boto3
timm
opencv-python-headless
httpx
python-dotenv
cloudinary
inference_sdk
